Make vehicle type enum publicly available
