+++ src/main/java/com/mappy/fpm/batches/tomtom/dbf/speedrestrictions/SpeedRestriction.java
@@ -1,6 +1,7 @@
 package com.mappy.fpm.batches.tomtom.dbf.speedrestrictions;
 
 import lombok.Data;
+import com.mappy.fpm.batches.tomtom.helpers.VehicleType;
 
 @Data
 public class SpeedRestriction {
@@ -14,24 +15,4 @@ public class SpeedRestriction {
         both, positive, negative
     }
 
-    public enum VehicleType {
-        all, passengerCars, residentialVehicles, taxi, publicBus;
-
-        public static VehicleType fromId(int id) {
-            switch(id) {
-                case 0:
-                    return all;
-                case 11:
-                    return passengerCars;
-                case 12:
-                    return residentialVehicles;
-                case 16:
-                    return taxi;
-                case 17:
-                    return publicBus;
-                default:
-                    throw new RuntimeException("Unknown vehicle type: " + id);
-            }
-        }
-    }
 }
