+++ src/main/java/com/mappy/fpm/batches/tomtom/helpers/VehicleType.java
@@ -1,0 +1,22 @@
+package com.mappy.fpm.batches.tomtom.helpers;
+
+public enum VehicleType {
+    all, passengerCars, residentialVehicles, taxi, publicBus;
+
+    public static VehicleType fromId(int id) {
+        switch(id) {
+            case 0:
+                return all;
+            case 11:
+                return passengerCars;
+            case 12:
+                return residentialVehicles;
+            case 16:
+                return taxi;
+            case 17:
+                return publicBus;
+            default:
+                throw new RuntimeException("Unknown vehicle type: " + id);
+        }
+    }
+}
